REFACTOR-#3622: Fix references to pandas*test
Fix references in Files to import modin.pandas.test or
using modin/pandas/test to new version

Also fix references in documentation.
