+++ setup.cfg
@@ -44,7 +44,7 @@ omit =
     # Skip tests
     modin/test/*
     modin/config/test/*
-    modin/pandas/test/*
+    modin/test/pandas/*
     modin/core/execution/dispatching/factories/test/*
     modin/experimental/cloud/test/*
     modin/experimental/pandas/test/*
